+++ src/constants.py
@@ -1,0 +1,3 @@
+# default pyethapp keystore path
+KEYSTORE_DIR_SUFFIX = ".config/pyethapp/keystore/"
+ROUND_DIGITS = 2