+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## [v20180418]
+
+  - Per account bet history, refs #16
+  - Handle unresolved bets, refs #51
+
+
 ## [v20180414]
 
   - UI testing, refs #10